lessons/lesson-05-reasoning-models/pyproject.toml
[project]
name = "transcript-analyzer"
version = "0.1.0"
description = "CLI Transcript Analyzer with AI-powered insights"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "google-generativeai>=0.3.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]